Added progress bar for upload
So I had to user lower level XMLHttpRequest in order to get track the progress of the API call rather than the regular fetch. Alternatively can try to use the boolean spinner that appears rather than a loading bar. The submit button still looks ugly, but I'm working on it.