Added cookies for upload and new logo
When a new user visits the webpage, generates a cookie with the attribute 'userId'. The userId is a random uuid value that represents the path to the user's folder in api/upload/user_files. This userId will be the folder path for all future uploads while the cookie persists in the browser. Also added a UserFiles.js client component which fetches and displays all files in that user's directory for the current cookie userId. You can test this behavior in the browser console > application > cookies. Also added dependencies for reading client-side cookies 'js-cookies' and generating random uuid's 'crypto'. Also fixed some CSS stylings that were deleted from a previous commit and removed some debug print statements.
TODO: add cookie link to download path, login auth, and debug for security.