Began merging python to front/backend
Main concern is that I can't run python scripts in a local host, so the MongoDB needs to be set up.

Python code has been slightly altered in order to accomodate for the js. It now takes one arg as an input, it being the userID

DownloadButton.js needs to be finished. I have put the location in which user files will be located at, so it is as simple as finding that specific file and outputting it

GenerateButton.js has prewritten code that should exec the python script, but it's something I cannot do until what was stated before

CalendarEvents.js should have the path to the ics file to displaym not sure if importing cookies and running ```userId = Cookies.get('userId');``` is enough to get the user id